refactored getStatusClass
